Only set subscribe property if operation is a subscription
